Fix non-returning agents with GCC
When taking adresses from references, GCC performs differently than
MSVC. This resulted in the agents not returning to their homes, because
the return callback bound a wrong adress. Now, I changed it to only use
pointers and everything seems to work.

Modified unit test move_location to test it. It worked with one agent in
GCC but not 5 or more.